refactor: keep the range-spec-style helpers out of the types package
`@pnpm/types` stays declaration-only: rangeSpecGranularity,
getRangeSpecStyle, and versionWithRangeSpecStyle move to
`@pnpm/pkg-manifest.utils`, next to createVersionSpecFromResolvedVersion,
which already renders manifest specifiers.
