+++ .changeset/save-prefix-equals.md
@@ -1,5 +1,5 @@
 ---
-"@pnpm/types": minor
+"@pnpm/pkg-manifest.utils": minor
 "@pnpm/installing.commands": patch
 "@pnpm/global.commands": patch
 "pnpm": minor
