+++ pnpm11/core/types/package.json
@@ -25,14 +25,12 @@
     "!*.map"
   ],
   "scripts": {
-    "test": "pn compile && pn .test",
-    ".test": "cross-env NODE_OPTIONS=\"$NODE_OPTIONS --experimental-vm-modules --disable-warning=ExperimentalWarning --disable-warning=DEP0169\" jest",
+    "test": "pn compile",
     "prepublishOnly": "tsgo --build",
     "compile": "tsgo --build && pn lint --fix",
-    "lint": "eslint \"src/**/*.ts\" \"test/**/*.ts\""
+    "lint": "eslint \"src/**/*.ts\""
   },
   "devDependencies": {
-    "@jest/globals": "catalog:",
     "@pnpm/types": "workspace:*"
   },
   "engines": {
