--- pnpm11/core/types/test/tsconfig.json
@@ -1,18 +1,0 @@
-{
-  "extends": "../tsconfig.json",
-  "compilerOptions": {
-    "noEmit": false,
-    "outDir": "../node_modules/.test.lib",
-    "rootDir": "..",
-    "isolatedModules": true
-  },
-  "include": [
-    "**/*.ts",
-    "../../../__typings__/**/*.d.ts"
-  ],
-  "references": [
-    {
-      "path": ".."
-    }
-  ]
-}
