+++ pnpm11/engine/pm/commands/package.json
@@ -51,6 +51,7 @@
     "@pnpm/lockfile.types": "workspace:*",
     "@pnpm/network.auth-header": "workspace:*",
     "@pnpm/os.env.path-extender": "catalog:",
+    "@pnpm/pkg-manifest.utils": "workspace:*",
     "@pnpm/resolving.npm-resolver": "workspace:*",
     "@pnpm/shell.path": "workspace:*",
     "@pnpm/store.connection-manager": "workspace:*",
