+++ pnpm11/engine/pm/commands/src/self-updater/selfUpdate.ts
@@ -12,9 +12,9 @@ import { createResolver, policyViolationToError, type ResolutionPolicyViolation 
 import { resolvePackageManagerIntegrities } from '@pnpm/installing.env-installer'
 import { readEnvLockfile } from '@pnpm/lockfile.fs'
 import { globalInfo, globalWarn } from '@pnpm/logger'
+import { versionWithRangeSpecStyle } from '@pnpm/pkg-manifest.utils'
 import { inferRangeSpecStyle, MINIMUM_RELEASE_AGE_VIOLATION_CODE } from '@pnpm/resolving.npm-resolver'
 import { createStoreController, type CreateStoreControllerOptions, shouldFetchFullMetadata } from '@pnpm/store.connection-manager'
-import { versionWithRangeSpecStyle } from '@pnpm/types'
 import { readProjectManifest } from '@pnpm/workspace.project-manifest-reader'
 import { isCI } from 'ci-info'
 import { pick } from 'ramda'
