+++ pnpm11/global/commands/tsconfig.json
@@ -49,6 +49,9 @@
       "path": "../../pkg-manifest/reader"
     },
     {
+      "path": "../../pkg-manifest/utils"
+    },
+    {
       "path": "../../store/connection-manager"
     },
     {
