+++ pnpm11/pkg-manifest/utils/src/index.ts
@@ -10,6 +10,7 @@ import { getSpecFromPackageManifest } from './getSpecFromPackageManifest.js'
 export * from './convertEnginesRuntimeToDependencies.js'
 export * from './getAllDependenciesFromManifest.js'
 export * from './getDependencyTypeFromManifest.js'
+export * from './rangeSpecStyle.js'
 export * from './updateProjectManifestObject.js'
 
 export { getAllUniqueSpecs, getSpecFromPackageManifest }
